+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - instead of dist specific workflows, use reusable ones (linter.yml, overtrue-phplint.yml, php-composer-dependencies.yml, phpcbf.yml)
 - VALIDATE_GITHUB_ACTIONS: false as it returns false positive `unknown Webhook event "workflow_call"` (todo reconsider later)
 - PHPStan level lowered to 8 due to stricter PHPStan:1.0
+- MyTableLister::bulkUpdateSQL passes $vars by value (it's not necessary to pass by reference)
 
 ### `Deprecated` for soon-to-be removed features